+++ package.json
@@ -32,7 +32,7 @@
     "react": "^18.3.1",
     "react-datepicker": "^7.5.0",
     "react-dom": "^18.3.1",
-    "react-router-dom": "^6.26.2",
+    "react-router-dom": "^6.28.0",
     "react-signature-canvas": "^1.0.6",
     "swiper": "^11.1.14",
     "tailwind-merge": "^2.5.4",
