+++ src/App.tsx
@@ -1,13 +1,14 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import { Login } from './pages/auth/Login';
-import { Signup } from './pages/auth/Signup';
-import { ProfileInput } from './pages/auth/ProfileInput';
-import { ProductCreatePage } from './pages/product/ProductCreatePage';
-import { ProductEditPage } from './pages/product/ProductEditPage';
-import { ProductViewPage } from './pages/product/ProductViewPage';
-import { ProfilePage } from './pages/profile/ProfilePage';
-import { ChatPage } from './pages/chat/ChatPage';
-import { MainPage } from './pages/main/MainPage';
+import MainPage from './pages/Main/page';
+import ProductCreatePage from './pages/Product/Create/page';
+import ProductEditPage from './pages/Product/Edit/page';
+import ProductViewPage from './pages/Product/View/page';
+import ChatPage from './pages/Chat/page';
+import LoginPage from './pages/Auth/Login/page';
+import SignupPage from './pages/Auth/Signup/page';
+import MyPage from './pages/Mypage/page';
+import OnboardingPage from './pages/Onboarding/page';
+
 
 function App() {
   return (
@@ -18,10 +19,10 @@ function App() {
         <Route path="/product/:id/edit" element={<ProductEditPage />} />
         <Route path="/product/:id" element={<ProductViewPage />} />
         <Route path="/chat" element={<ChatPage />} />
-        <Route path="/login" element={<Login />} />
-        <Route path="/signup" element={<Signup />} />
-        <Route path="/signup/profile" element={<ProfileInput />} />
-        <Route path="/profile" element={<ProfilePage />} />
+        <Route path="/login" element={<LoginPage />} />
+        <Route path="/signup" element={<SignupPage />} />
+        <Route path="/signup/profile" element={<OnboardingPage />} />
+        <Route path="/profile" element={<MyPage />} />
       </Routes>
     </Router>
   );
